Keras Neural Network architecture base class

A KerasArchitecture is a parametrized abstraction of a keras model. It can be used to more easily and flexibly add architectures.

Initialize architecture

Obtain the model. Called by Learner during train_internal.

Setter method for 'x_transform' and 'y_transform'.

This is an architecture used for custom, user-supplied models. The model, i.e a compiled keras model is supplied to the learner as a hyperparameter.

Arguments

y_transform
function
Function used to transform the targets for the model.

param_set
ParamSet
A methods / architecture's ParamSet.

build_arch_fun
Task
Function that instantiates and compiles a model.

x_transform
list
Function used to transform the data for the model.

name
character Either 'x' or 'y'.

transform
function
Function to set for the architecture

Fields

param_set

ParamSet
A methods \ architecture's ParamSet.

build_arch_fun

function
Function that instantiates and compiles a model.

transforms

list
The corresponding x- and y_transform.