Keras Feed Forward Neural Network

Feed Forward Neural Network using Keras and Tensorflow. This learner builds and compiles the keras model from the hyperparameters in param_set, and does not require a supplied and compiled model.

Calls keras::fit from package keras. Layers are set up as follows:

The inputs are connected to a layer_dropout, applying the input_dropout. Afterwards, each layer_dense() is followed by a layer_activation, and depending on hyperparameters by a layer_batch_normalization and or a layer_dropout depending on the architecture hyperparameters. This is repeated length(layer_units) times, i.e. one 'dense->activation->batchnorm->dropout' block is appended for each layer_unit.

Parameters:
Most of the parameters can be obtained from the keras documentation. Some exceptions are documented here.

layer_units: An integer vector storing the number of units in each consecutive layer. layer_units = c(32L, 32L, 32L) results in a 3 layer network with 32 neurons in each layer. Can be integer(0), in which case we fit a (multinomial) logistic regression model.

initializer: Weight and bias initializer.

"glorot_uniform"  : initializer_glorot_uniform(seed)
"glorot_normal"   : initializer_glorot_normal(seed)
"he_uniform"      : initializer_he_uniform(seed)
"..."             : see `??keras::initializer`

optimizer: Some optimizers and their arguments can be found below.
Inherits from tensorflow.python.keras.optimizer_v2.

"sgd"     : optimizer_sgd(lr, momentum, decay = decay),
"rmsprop" : optimizer_rmsprop(lr, rho, decay = decay),
"adagrad" : optimizer_adagrad(lr, decay = decay),
"adam"    : optimizer_adam(lr, beta_1, beta_2, decay = decay),
"nadam"   : optimizer_nadam(lr, beta_1, beta_2, schedule_decay = decay)

regularizer: Regularizer for keras layers:

"l1"      : regularizer_l1(l = 0.01)
"l2"      : regularizer_l2(l = 0.01)
"l1_l2"   : regularizer_l1_l2(l1 = 0.01, l2 = 0.01)

class_weights: needs to be a named list of class-weights for the different classes numbered from 0 to c-1 (for c classes).

Example:
wts = c(0.5, 1)
setNames(as.list(wts), seq_len(length(wts)) - 1)

callbacks: A list of keras callbacks. See ?callbacks.

Format

R6::R6Class() inheriting from mlr3::LearnerClassif.

Construction

LearnerClassifKerasFF$new()
mlr3::mlr_learners$get("classif.kerasff")
mlr3::lrn("classif.kerasff")

See also

Examples

learner = mlr3::lrn("classif.kerasff")
print(learner)

#> <LearnerClassifKerasFF:classif.kerasff>
#> * Model: -
#> * Parameters: epochs=30, activation=relu, layer_units=32,32,32,
#>   initializer=<tensorflow.python.ops.init_ops_v2.GlorotUniform>,
#>   optimizer=<keras.optimizer_v2.adam.Adam>,
#>   regularizer=<keras.regularizers.L1L2>, use_batchnorm=FALSE,
#>   use_dropout=FALSE, dropout=0, input_dropout=0, callbacks=<list>,
#>   loss=categorical_crossentropy, validation_split=0.3333,
#>   batch_size=128
#> * Packages: keras
#> * Predict Type: response
#> * Feature types: integer, numeric
#> * Properties: multiclass, twoclass

# available parameters:
learner$param_set$ids()

#>  [1] "epochs"           "layer_units"      "initializer"      "regularizer"     
#>  [5] "optimizer"        "activation"       "use_batchnorm"    "use_dropout"     
#>  [9] "dropout"          "input_dropout"    "class_weights"    "validation_split"
#> [13] "batch_size"       "callbacks"        "verbose"          "loss"